[backend] remove sorting dependency cycle (#11149)
