+++ opencti-platform/opencti-graphql/src/modules/pir/pir-domain.ts
@@ -46,7 +46,8 @@ import { deleteInternalObject } from '../../domain/internalObject';
 import { registerConnectorForPir, unregisterConnectorForIngestion } from '../../domain/connector';
 import type { BasicStoreCommon, BasicStoreObject } from '../../types/store';
 import { RELATION_OBJECT } from '../../schema/stixRefRelationship';
-import { checkEEAndPirAccess, createPirRelation, serializePir, updatePirExplanations } from './pir-utils';
+import { createPirRelation, serializePir, updatePirExplanations } from './pir-utils';
+import { checkEEAndPirAccess } from './pir-checkPirAccess';
 import { ForbiddenAccess, FunctionalError } from '../../config/errors';
 import { ABSTRACT_STIX_REF_RELATIONSHIP, ENTITY_TYPE_CONTAINER } from '../../schema/general';
 import { extractFilterKeyValues } from '../../utils/filtering/filtering-utils';
