+++ opencti-platform/opencti-graphql/src/modules/pir/pir-utils.ts
@@ -27,24 +27,6 @@ import { RELATION_FROM_TYPES_FILTER } from '../../utils/filtering/filtering-cons
 import { elUpdate } from '../../database/engine';
 import { INDEX_STIX_DOMAIN_OBJECTS } from '../../database/utils';
 import type { BasicStoreEntity } from '../../types/store';
-import { checkEnterpriseEdition } from '../../enterprise-edition/ee';
-import { getEntityFromIdFromCache } from '../../database/cache';
-
-/**
- * Helper function to check a user has access to a pir functionnalities
- */
-export const checkEEAndPirAccess = async (context: AuthContext, user: AuthUser, pirId?: string | null) => {
-  // check EE
-  await checkEnterpriseEdition(context);
-  // check user has access to the PIR
-  if (!pirId) {
-    throw FunctionalError('No Pir ID provided');
-  }
-  const pir = await getEntityFromIdFromCache(context, user, pirId, ENTITY_TYPE_PIR);
-  if (!pir) {
-    throw FunctionalError('No PIR found', { pirId });
-  }
-};
 
 /**
  * Helper function to parse filters that are saved as string in elastic.
